install_script

+++ install_webserver.sh
@@ -7,8 +7,13 @@ sudo systemctl start apache2
 
 # Move the website inside /var/www/html/
 
-# Clear the directory
-sudo rm /var/www/html/*
+# Clear the directory if empty
+if [ -z "$(ls -A /var/www/html)" ]; then
+   echo "/var/www/html/ is empty"
+else
+   sudo rm /var/www/html/*
+fi
+
 
 # Copy the build folder insides to the /var/www/html
 sudo mv ./build/* /var/www/html/
